refactor(tests): move RAG tests into tests/rag/ subdirectory
Mirror the src/okp_mcp/rag/ layout so RAG tests live in their own
subpackage with dedicated fixtures instead of sharing the root conftest.
