+++ tests/conftest.py
@@ -6,6 +6,8 @@ import respx
 
 from okp_mcp.config import ServerConfig
 
+# RAG-specific fixtures live in tests/rag/conftest.py.
+
 
 @pytest.fixture
 def sample_solr_response():
@@ -44,35 +46,6 @@ def solr_mock(sample_solr_response):
     with respx.mock:
         route = respx.get(config.solr_endpoint).mock(return_value=httpx.Response(200, json=sample_solr_response))
         yield route
-
-
-@pytest.fixture
-def rag_chunk_response():
-    """Realistic portal-rag chunk response shared by lexical, hybrid, and semantic tests."""
-    return {
-        "response": {
-            "numFound": 1,
-            "docs": [
-                {
-                    "doc_id": "/security/cve/CVE-2024-42225_chunk_2",
-                    "parent_id": "/security/cve/CVE-2024-42225",
-                    "title": "CVE-2024-42225 - Red Hat Customer Portal",
-                    "chunk": "A potential flaw was found...",
-                    "headings": "CVE-2024-42225,Description",
-                    "chunk_index": 2,
-                    "num_tokens": 49,
-                    "is_chunk": True,
-                }
-            ],
-        }
-    }
-
-
-@pytest.fixture
-async def rag_client():
-    """Async httpx client with automatic cleanup for RAG tests."""
-    async with httpx.AsyncClient() as client:
-        yield client
 
 
 def pytest_collection_modifyitems(config: pytest.Config, items: list[pytest.Item]) -> None:
