+++ tests/rag/test_lexical.py
